+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ExpressionSuggesterSpec.scala
@@ -323,9 +323,9 @@ class ExpressionSuggesterSpec
   test("should suggest global meta variable") {
     spelSuggestionsFor("#meta.") shouldBe List(
       ExpressionSuggestion("empty", Typed[Boolean], fromClass = true, None, Nil),
-      suggestion("processLabels", Typed.genericTypeClass[java.util.List[String]](List(Typed[String]))),
       suggestion("processName", Typed[String]),
       suggestion("properties", Typed.record(Map("scenarioProperty" -> Typed[String]))),
+      suggestion("scenarioLabels", Typed.genericTypeClass[java.util.List[String]](List(Typed[String]))),
     )
   }